requirements.txt
faker>=37.4.0
requests>=2.31.0
pytest>=7.0.0
allure-pytest>=2.13.0
allure-python-commons>=2.13.0 